Move UseQuery rtk type to metabase/api/types
