+++ frontend/src/metabase/plugins/oss/library.ts
@@ -2,12 +2,12 @@ import type { BaseQueryFn, QueryDefinition } from "@reduxjs/toolkit/query";
 import type { ComponentType, ReactNode } from "react";
 
 import type { TagType } from "metabase/api/tags";
+import type { UseQuery } from "metabase/api/types/rtk";
 import type {
   OmniPickerCollectionItem,
   OmniPickerItem,
 } from "metabase/common/components/Pickers/EntityPicker/types";
 import type { MiniPickerCollectionFolderItem } from "metabase/common/components/Pickers/MiniPicker/types";
-import type { UseQuery } from "metabase/entities/containers/rtk-query/types/rtk";
 import { PluginPlaceholder } from "metabase/plugins/components/PluginPlaceholder";
 import type {
   Collection,
