+++ frontend/src/metabase/plugins/oss/remote-sync.ts
@@ -6,9 +6,9 @@ import type {
 import type { ComponentType, ReactNode } from "react";
 
 import type { TagType } from "metabase/api/tags";
+import type { UseQuery } from "metabase/api/types/rtk";
 import type { CollectionTreeItem } from "metabase/collections/utils";
 import type { ITreeNodeItem } from "metabase/common/components/tree/types";
-import type { UseQuery } from "metabase/entities/containers/rtk-query/types/rtk";
 import type {
   GitSyncSetupMenuItemProps,
   SyncedCollectionsSidebarSectionProps,
